// src/notification/notification.enum.ts
export enum Notification_type {
  quote_sent = 'quote sent',
  quote_accepted = 'quote accepted',
  task_finished = 'task finished',
  task_assigned = 'task assigned',
  task_unassigned = 'task unassigned',
  quote_received = 'quote received',
  event_finished = 'event finished',
  event_cancelled = 'event cancelled',
}